backend/requirements.txt
fastapi<0.103
uvicorn
python-dotenv
supabase
pydantic<2
reportlab
